+++ src/endstone/core/block/biome.h
@@ -1,0 +1,33 @@
+// Copyright (c) 2024, The Endstone Project. (https://endstone.dev) All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include "bedrock/world/level/biome/biome.h"
+#include "endstone/block/biome.h"
+
+namespace endstone::core {
+
+class EndstoneBiome : public endstone::Biome {
+public:
+    explicit EndstoneBiome(const ::Biome &biome);
+    [[nodiscard]] BiomeId getId() const override;
+    [[nodiscard]] std::string getTranslationKey() const override;
+    [[nodiscard]] const ::Biome &getHandle() const;
+
+private:
+    const ::Biome &biome_;
+};
+
+}  // namespace endstone::core