+++ src/endstone/core/block/block_snapshot.cpp
@@ -14,100 +14,51 @@
 
 #include "endstone/core/block/block_snapshot.h"
 
-#include "endstone/core/block/block.h"
+#include "bedrock/world/level/block/block.h"
 #include "endstone/core/block/block_data.h"
-#include "endstone/core/block/block_face.h"
+#include "endstone/core/block/block_state.h"
+#include "endstone/core/server.h"
 
 namespace endstone::core {
 
-EndstoneBlockSnapshot::EndstoneBlockSnapshot(std::unique_ptr<EndstoneBlockState> state)
-    : state_(std::move(state))
-{
-}
-
 const BlockType &EndstoneBlockSnapshot::getType() const
 {
-    return state_->getType();
+    return *BlockType::get(placed_->getName().getString());
 }
 
 void EndstoneBlockSnapshot::setType(BlockTypeId type)
 {
-    state_->setType(type);
+    setData(*EndstoneServer::getInstance().createBlockData(type));
 }
 
 void EndstoneBlockSnapshot::setType(BlockTypeId type, bool /*apply_physics*/)
 {
-    state_->setType(type);
+    setType(type);
 }
 
 std::unique_ptr<BlockData> EndstoneBlockSnapshot::getData() const
 {
-    return state_->getData();
+    return std::make_unique<EndstoneBlockData>(*placed_);
 }
 
 void EndstoneBlockSnapshot::setData(const BlockData &data)
 {
-    state_->setData(data);
+    placed_ = &static_cast<const EndstoneBlockData &>(data).getHandle();
 }
 
 void EndstoneBlockSnapshot::setData(const BlockData &data, bool /*apply_physics*/)
 {
-    state_->setData(data);
-}
-
-std::unique_ptr<Block> EndstoneBlockSnapshot::getRelative(int offset_x, int offset_y, int offset_z)
-{
-    return getDimension().getBlockAt(getX() + offset_x, getY() + offset_y, getZ() + offset_z);
-}
-
-std::unique_ptr<Block> EndstoneBlockSnapshot::getRelative(BlockFace face)
-{
-    return getRelative(face, 1);
-}
-
-std::unique_ptr<Block> EndstoneBlockSnapshot::getRelative(BlockFace face, int distance)
-{
-    return getRelative(EndstoneBlockFace::getOffsetX(face) * distance, EndstoneBlockFace::getOffsetY(face) * distance,
-                       EndstoneBlockFace::getOffsetZ(face) * distance);
-}
-
-Dimension &EndstoneBlockSnapshot::getDimension() const
-{
-    return state_->getDimension();
-}
-
-int EndstoneBlockSnapshot::getX() const
-{
-    return state_->getX();
-}
-
-int EndstoneBlockSnapshot::getY() const
-{
-    return state_->getY();
-}
-
-int EndstoneBlockSnapshot::getZ() const
-{
-    return state_->getZ();
-}
-
-Location EndstoneBlockSnapshot::getLocation() const
-{
-    return state_->getLocation();
+    setData(data);
 }
 
 std::unique_ptr<BlockState> EndstoneBlockSnapshot::captureState() const
 {
-    const auto &block = static_cast<const EndstoneBlockData &>(*state_->getData()).getHandle();
-    return std::make_unique<EndstoneBlockState>(state_->getDimension(), BlockPos(getX(), getY(), getZ()), block);
+    return std::make_unique<EndstoneBlockState>(getDimension(), getPosition(), *placed_);
 }
 
 std::unique_ptr<Block> EndstoneBlockSnapshot::clone() const
 {
-    const auto &block = static_cast<const EndstoneBlockData &>(*state_->getData()).getHandle();
-    auto cloned_state =
-        std::make_unique<EndstoneBlockState>(state_->getDimension(), BlockPos(getX(), getY(), getZ()), block);
-    return std::make_unique<EndstoneBlockSnapshot>(std::move(cloned_state));
+    return std::make_unique<EndstoneBlockSnapshot>(block_source_, block_pos_, *placed_);
 }
 
 }  // namespace endstone::core